Added the /tpoll command
This will allow the user to create a poll with all the tiers on it already with out having to type every single tier or meta game in a single poll, this is more efficient  and takes like 2 seconds a very useful tool.

added /tpoll command

This poll will allow the user to create a tier poll with having to write the tiers in a single poll over and over, it's efficient and easy, a very useful tool.

Added the /tpoll command

This will allow the user to create a poll with all the tiers on it already with out having to type every single tier or meta game in a single poll, this is more efficient  and takes like 2 seconds a very useful tool.

added /tpoll command

This poll will allow the user to create a tier poll with having to write the tiers in a single poll over and over, it's efficient and easy, a very useful tool.

Cleanup