Update pokedex.js
Instead of guess work stats Digimon Dawn and Dusk are the basis for these stats. This is a Nintendo DS game, it covered everyone but Elecmon, Meramon, Punimon and Yuramon. So I did a educated guess on that based on other stats in the same tier. Elecmon Stats are leomons stats -33% because it looked good and they share a evo line. Meramon runs close to Leomon but is a bit different, I figured since they both have a human like build it would be a okay guess.